Workflow trigger used in ResourceFunction plural fix
If a ResourceFunction uses the same resource type as a Workflow trigger,
the watch can't be started. That is because kr8s uses APIObject
subclasses in order to maintain a cache of known resource types and we
use a placeholder to indicate the plural is unknown. This adds a check
to ensure that the plural placeholder triggers correcting the
subclasses.